fix request event occurrence: remove requests when occurrence is removed; verify if event occurrence exists before try to change it

+++ src/protected/application/lib/MapasCulturais/Entities/RequestEventOccurrence.php
@@ -34,13 +34,17 @@ class RequestEventOccurrence extends Request{
 
     function _doApproveAction() {
         $eo = $this->getEventOccurrence();
-        $eo->status = EventOccurrence::STATUS_ENABLED;
-        $eo->save(true);
+        if($eo){
+            $eo->status = EventOccurrence::STATUS_ENABLED;
+            $eo->save(true);
+        }
     }
 
     function _doRejectAction() {
         $eo = $this->getEventOccurrence();
-        $eo->delete(true);
+        if($eo){
+            $eo->delete(true);
+        }
         parent::_doRejectAction();
     }
 }
